+++ apps/catalogue/app/pages/[catalogue]/[resourceType]/[resource]/index.vue
@@ -2,6 +2,7 @@
 import { useFetch, useHead, useRoute, useRuntimeConfig } from "#app";
 import {
   logError,
+  moduleToString,
   removeChildIfParentSelected,
   useDatasetStore,
 } from "#imports";
@@ -38,12 +39,14 @@ import LayoutsDetailPage from "../../../../components/layouts/DetailPage.vue";
 import ReferenceCard from "../../../../components/ReferenceCard.vue";
 import ReferenceCardList from "../../../../components/ReferenceCardList.vue";
 import SideNavigation from "../../../../components/SideNavigation.vue";
+import CardButton from "../../../../components/store/CartButton.vue";
 import SubpopulationDisplay from "../../../../components/SubpopulationDisplay.vue";
 import TableContent from "../../../../components/table/Content.vue";
 import VariableDisplay from "../../../../components/VariableDisplay.vue";
 import collectionEventsQuery from "../../../../gql/collectionEvents";
 import datasetQuery from "../../../../gql/datasets";
-import { resourceQuery } from "../../../../gql/resource";
+import fileFragment from "../../../../gql/fragments/file";
+import ontologyFragment from "../../../../gql/fragments/ontology";
 import subpopulationsQuery from "../../../../gql/subpopulations";
 import variablesQuery from "../../../../gql/variables";
 import dateUtils from "../../../../utils/dateUtils";
@@ -52,10 +55,192 @@ import { getKey } from "../../../../utils/variableUtils";
 
 const config = useRuntimeConfig();
 const route = useRoute();
-const schema = config.public.schema as string;
+const schema = config.public.schema;
 const datasetStore = useDatasetStore();
 
 const variables = { id: route.params.resource };
+
+const resourceQuery = `
+  query Resources($id: String) {
+    Resources(filter: { id: { equals: [$id] } }) {
+      id
+      pid
+      acronym
+      name
+      description
+      website
+      logo {
+        url
+      }
+      contactEmail
+      type {
+        name
+      }
+      cohortType {
+        name
+      }
+      registryOrHealthRecordType {
+        name
+      }
+      networkType {
+        name
+      }
+      clinicalStudyType {
+        name
+      }
+      keywords
+      externalIdentifiers {
+        identifier
+        externalIdentifierType{name}
+      }
+      populationAgeGroups {
+        name order code parent { code }
+      }
+      dateLastRefresh
+      startYear
+      endYear
+      continents {
+        name
+        order
+      }
+      countries {
+        name order
+      }
+      regions {
+        name
+        order
+      }
+      numberOfParticipants
+      numberOfParticipantsWithSamples
+      designDescription
+      designSchematic ${moduleToString(fileFragment)}
+      design {
+        definition
+        name
+      }
+      dataCollectionType {
+        definition
+        name
+      }
+      dataCollectionDescription
+      reasonSustained
+      unitOfObservation
+      recordTrigger
+      populationOncologyTopology ${moduleToString(ontologyFragment)}
+      populationOncologyMorphology ${moduleToString(ontologyFragment)}
+      inclusionCriteria ${moduleToString(ontologyFragment)}
+      otherInclusionCriteria
+      exclusionCriteria ${moduleToString(ontologyFragment)}
+      otherExclusionCriteria
+      publications(orderby: {title:ASC}) {
+        doi
+        title
+        isDesignPublication
+      }
+      collectionEvents {
+        name
+        description
+        startDate
+        endDate
+        numberOfParticipants
+        ageGroups ${moduleToString(ontologyFragment)}
+        dataCategories ${moduleToString(ontologyFragment)}
+        sampleCategories ${moduleToString(ontologyFragment)}
+        areasOfInformation ${moduleToString(ontologyFragment)}
+        subpopulations {
+          name
+        }
+        coreVariables
+      }
+      peopleInvolved {
+        roleDescription
+        firstName
+        lastName
+        prefix
+        initials
+        email
+        title {
+          name
+        }
+        organisation {
+          name
+          website
+          organisationWebsite
+          email
+          organisation {
+            name
+            label
+          }
+      }
+        role ${moduleToString(ontologyFragment)}
+      }
+      organisationsInvolved(orderby: {name: ASC})  {
+        id
+        name
+        website
+        isLeadOrganisation
+        role ${moduleToString(ontologyFragment)}
+        otherOrganisation
+        organisation {
+          name
+          acronym
+          website
+          country {
+            name
+            order
+          }
+        }
+      }
+      subpopulations {
+          name
+          mainMedicalCondition ${moduleToString(ontologyFragment)}
+      }
+      dataAccessConditions ${moduleToString(ontologyFragment)}
+      dataAccessConditionsDescription
+      dataUseConditions ${moduleToString(ontologyFragment)}
+      dataAccessFee
+      releaseType ${moduleToString(ontologyFragment)}
+      releaseDescription
+      fundingStatement
+      acknowledgements
+      linkageOptions
+      prelinked
+      documentation {
+        name
+        description
+        url
+        file ${moduleToString(fileFragment)}
+      }
+      datasets {
+        name
+      }
+      partOfNetworks {
+        id
+        name
+        description
+        website
+        logo {
+          url
+        }
+        catalogueType {
+          name
+        }
+      }
+      publications_agg {
+        count
+      }
+      subpopulations_agg {
+        count
+      }
+      collectionEvents_agg{
+        count
+      }
+    }
+    Variables_agg(filter: { resource: { id: { equals: [$id] } } }) {
+      count
+    }
+  }
+`;
 
 interface IResourceQueryResponseValue extends IResources {
   publications_agg?: { count: number };
@@ -111,10 +296,6 @@ const subpopulationCount = computed(
 );
 
 const variableCount = computed(() => data.value?.data?.Variables_agg.count);
-
-const isInShoppingCart = computed(() =>
-  resource.value?.id ? datasetStore.resourceIsInCart(resource.value?.id) : false
-);
 
 function collectionEventMapper(item: any) {
   return {
@@ -534,16 +715,6 @@ const showPopulation = computed(
       (item) => item.content !== undefined && item.content !== ""
     ).length
 );
-
-function onShoppingCartInput() {
-  if (resource.value) {
-    if (isInShoppingCart.value) {
-      datasetStore.removeFromCart(resource.value.id);
-    } else {
-      datasetStore.addToCart(resource.value);
-    }
-  }
-}
 </script>
 <template>
   <LayoutsDetailPage>
@@ -564,25 +735,12 @@ function onShoppingCartInput() {
           <BreadCrumbs :crumbs="crumbs" />
         </template>
         <template #title-suffix>
-          <label
-            v-if="datasetStore.isEnabled"
-            :for="`${resource?.id}-shopping-cart-input`"
-            class="xl:flex xl:justify-end px-2 py-1 rounded-3px cursor-pointer text-link hover:text-blue-800 focus:text-blue-800"
-            :class="{
-              'items-baseline xl:items-center mt-0.5 xl:mt-0': true,
-              'bg-blue-500 text-white hover:text-white': isInShoppingCart,
-            }"
-          >
-            <BaseIcon name="shopping-cart-add" :width="21" />
-            <span class="sr-only"></span>
-            <input
-              type="checkbox"
-              :id="`${resource?.id}-shopping-cart-input`"
-              class="sr-only"
-              :modelValue="isInShoppingCart"
-              @input="onShoppingCartInput"
-            />
-          </label>
+          <CardButton
+            v-if="resource && datasetStore.isEnabled"
+            :resource="resource"
+            :compact="false"
+            :invert="true"
+          />
         </template>
       </PageHeader>
     </template>
